fix: enforce strict collections output validation

+++ src/schemas/collections.ts
@@ -4,7 +4,7 @@ export const AccountActionSchema = z.object({
   customer: z.string(),
   invoice_id: z.string(),
   risk_tier: z.enum(["low", "medium", "high"]),
-  priority_score: z.number().min(0).max(100),
+  priority_score: z.number().int().min(0).max(100),
   suggested_action: z.string(),
   escalate: z.boolean(),
   email_draft: z.string(),
